+++ R/link-href.R
@@ -128,8 +128,8 @@ href_topic_remote <- function(topic, package) {
   if (!is.null(reference_url)) {
     paste0(reference_url, paste0("/", rdname, ".html"))
   } else {
-    # Fall back to rdocumentation.org which almost certainly works
-    paste0("https://www.rdocumentation.org/packages/", package, "/topics/", rdname)
+    # Fall back to rdrr.io
+    paste0("https://rdrr.io/pkg/", package, "/man/", rdname, ".html")
   }
 }