+++ R/build.r
@@ -264,13 +264,14 @@
 #' }
 build_site <- function(pkg = ".",
                        examples = TRUE,
-                       document = FALSE,
+                       document = devel,
                        run_dont_run = FALSE,
                        seed = 1014,
                        lazy = FALSE,
                        override = list(),
                        preview = NA,
-                       new_process = TRUE) {
+                       new_process = TRUE,
+                       devel = TRUE) {
 
   if (new_process) {
     build_site_external(
@@ -281,7 +282,8 @@ build_site <- function(pkg = ".",
       seed = seed,
       lazy = lazy,
       override = override,
-      preview = preview
+      preview = preview,
+      devel = devel
     )
   } else {
     build_site_local(
@@ -292,7 +294,8 @@ build_site <- function(pkg = ".",
       seed = seed,
       lazy = lazy,
       override = override,
-      preview = preview
+      preview = preview,
+      devel = devel
     )
   }
 }
@@ -304,7 +307,8 @@ build_site_external <- function(pkg = ".",
                                 seed = 1014,
                                 lazy = FALSE,
                                 override = list(),
-                                preview = NA) {
+                                preview = NA,
+                                devel = TRUE) {
   args <- list(
     pkg = pkg,
     examples = examples,
@@ -315,6 +319,7 @@ build_site_external <- function(pkg = ".",
     override = override,
     preview = FALSE,
     new_process = FALSE,
+    devel = devel,
     crayon_enabled = crayon::has_color(),
     crayon_colors = crayon::num_colors(),
     pkgdown_internet = has_internet()
@@ -344,7 +349,8 @@ build_site_local <- function(pkg = ".",
                        seed = 1014,
                        lazy = FALSE,
                        override = list(),
-                       preview = NA
+                       preview = NA,
+                       devel = TRUE
                        ) {
 
   pkg <- section_init(pkg, depth = 0, override = override)
@@ -355,7 +361,7 @@ build_site_local <- function(pkg = ".",
 
   init_site(pkg)
 
-  build_home(pkg, override = override, preview = FALSE)
+  build_home(pkg, override = override, preview = FALSE, knit = devel)
   build_reference(pkg,
     lazy = lazy,
     document = document,
@@ -363,7 +369,8 @@ build_site_local <- function(pkg = ".",
     run_dont_run = run_dont_run,
     seed = seed,
     override = override,
-    preview = FALSE
+    preview = FALSE,
+    devel = devel
   )
   build_articles(pkg, lazy = lazy, override = override, preview = FALSE)
   build_tutorials(pkg, override = override, preview = FALSE)