+++ R/usage.R
@@ -112,7 +112,7 @@ fun_info <- function(fun) {
       list(
         type = "s4",
         name = as.character(x[[2]]),
-        signature = purrr::map_chr(as.list(x[[3]][-1]), as.character)
+        signature = as.character(purrr::pluck(x[[3]][-1]))
       )
     } else if (is_call(x, c("::", ":::"))) {
       # TRUE if fun has a namespace, pkg::fun()