+++ tests/testthat/test-figure.R
@@ -11,6 +11,6 @@ test_that("can override defaults in _pkgdown.yml", {
   expect_setequal(img, c("figure-1.jpg", "figure-2.jpg"))
 
   expect_output(build_articles(figure))
-  img <- path_file(dir_ls(path(figure, "docs", "articles"), glob = "*.jpg", recursive = TRUE))
+  img <- path_file(dir_ls(path(figure, "docs", "articles"), glob = "*.jpg", recurse = TRUE))
   expect_equal(img, "unnamed-chunk-1-1.jpg")
 })