+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * One can override the title of the homepage via a `title` field in the config (#957, @maelle).
 
+* Links to external documentation now point to [rdrr.io](https://rdrr.io) (#998).
+
 * A default 404 page (`404.html`) is built from content in `.github/404.md` (#947).
 
 * Updated fontawesome to 5.7.1. [fontawesome 5 deprecated the `fa` prefix style](https://fontawesome.com/how-to-use/on-the-web/referencing-icons/basic-use),